[ADD] extract pharmacies list to separate page
